Spelling and grammar suggested changes
I suggest these changes are made to the spelling and grammar of a few
portions of Linux Sea. Please realize I'm coming from a US English
understanding of spelling/grammar. I also fixed a couple of references
to make.conf being in /etc/ instead of /etc/portage/ so if the grammar
changes are rejected at least those may be helpful. :)